ShelterLuv photos sync (#434)
* modified:   .github/workflows/deploy.yml

* modified:   .github/workflows/deploy.yml

* modified:   .github/workflows/deploy.yml

* modified:   Cloud Functions/asm/main.py
	modified:   Cloud Functions/shelterluv/main.py

* Delete .github/workflows/deploy.yml

* modified:  update photos sync

* modified:   Flutter/shelter_partner/lib/models/photo.dart
	modified:   Flutter/shelter_partner/lib/repositories/add_note_repository.dart
	modified:   Flutter/shelter_partner/lib/repositories/edit_animal_repository.dart

+++ Flutter/shelter_partner/lib/repositories/edit_animal_repository.dart
@@ -2,10 +2,18 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import '../models/photo.dart';
+
 class EditAnimalRepository {
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
 
-  Future<void> deleteItem(String shelterId, String animalType, String animalId, String field, String itemId) async {
+  Future<void> deleteItem(
+    String shelterId,
+    String animalType,
+    String animalId,
+    String field,
+    String itemId,
+  ) async {
     final documentRef = _firestore
         .collection('shelters')
         .doc(shelterId)
@@ -20,7 +28,8 @@ class EditAnimalRepository {
         final List<dynamic> items = data?[field] ?? [];
 
         // Filter out the item with the matching id
-        final updatedItems = items.where((item) => item['id'] != itemId).toList();
+        final updatedItems =
+            items.where((item) => item['id'] != itemId).toList();
 
         // Update the field with the filtered list
         transaction.update(documentRef, {field: updatedItems});
@@ -28,23 +37,55 @@ class EditAnimalRepository {
     });
   }
 
-  Future<void> deletePhotoFromStorage(String shelterId, String animalId, String photoId) async {
-
+  Future<void> deletePhotoFromStorage(
+    String shelterId,
+    String animalId,
+    Photo photo, // PhotoId to Photo object
+    String animalType,
+  ) async {
     final storage = FirebaseStorage.instance;
     final List<String> sizes = ['100x100', '250x250', '500x500', '750x750'];
 
-    // Delete the original photo
-    final originalPhotoRef = storage.ref().child('$shelterId/$animalId/$photoId');
-    await originalPhotoRef.delete();
+    final animalRef = _firestore
+        .collection('shelters')
+        .doc(shelterId)
+        .collection(animalType == "dog" ? "dogs" : "cats")
+        .doc(animalId);
 
-    // Delete the resized photos
-    for (final size in sizes) {
-      final resizedPhotoRef = storage.ref().child('$shelterId/$animalId/${photoId}_$size');
+    // For manual photos , delete from storage
+    if (photo.source == 'manual') {
       try {
-        await resizedPhotoRef.delete();
+        // Delete original photo
+        final originalPhotoRef = storage.ref().child(
+          '$shelterId/$animalId/${photo.id}',
+        );
+        await originalPhotoRef.delete();
+
+        // Delete resized photos
+        for (final size in sizes) {
+          final resizedPhotoRef = storage.ref().child(
+            '$shelterId/$animalId/${photo.id}_$size',
+          );
+          try {
+            await resizedPhotoRef.delete();
+          } catch (e) {
+            print('Failed to delte resized photos: ${e.toString()}');
+          }
+        }
       } catch (e) {
-        print('Failed to delete resized photo: ${e.toString()}');
-        // You can handle the error if needed
+        print('Error deleting manual photo from storage: $e');
+      }
+    }
+    // For ShelterLuv/asm photos , add to deleted_photos collection
+    else if (photo.source == 'shelterluv' || photo.source == 'asm') {
+      try {
+        await animalRef.collection('deleted_photos').add({
+          'url': photo.url,
+          'deletedAt': Timestamp.now(),
+          'source': photo.source,
+        });
+      } catch (e) {
+        print('Error adding to deleted photos: $e');
       }
     }
   }
